FDG-9776 update download wrapper, add unit test page
