+++ src/components/download-wrapper/data-dictionary-download-helper.js
@@ -50,11 +50,6 @@ export const convertDataDictionaryToCsv = dataset => {
   }
 };
 
-export const replaceNbsps = str => {
-  const re = new RegExp(String.fromCharCode(160), 'g'); // global search for " "
-  return str.replace(re, ' ');
-};
-
 const addDataset = (fields, dataset) => {
   return fields.map(field => Object.assign({ dataset: dataset.name }, field));
 };
