+++ src/components/download-wrapper/download-wrapper.jsx
@@ -3,8 +3,7 @@ import { wrapper, downloadDescription, describer, dateStringStyle } from './down
 import Truncator from '../truncate/truncate';
 import DownloadItemButton from './download-item-button/download-item-button';
 import Analytics from '../../utils/analytics/analytics';
-import { triggerDataDictionaryDownload } from './download-helpers';
-import { calcDictionaryDownloadSize, convertDataDictionaryToCsv } from './data-dictionary-download-helper';
+import { calcDictionaryDownloadSize, convertDataDictionaryToCsv, triggerDataDictionaryDownload } from './data-dictionary-download-helper';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faFileDownload, faSpinner } from '@fortawesome/free-solid-svg-icons';
 import DownloadToggle from './download-toggle/download-toggle';
